Mark the backup and the glyph API as unstable.
Also, split out some internals from allegro_font.h into an internal header, so
we don't have to install aintern_list.h.
