+++ addons/font/allegro5/allegro_font.h
@@ -2,7 +2,6 @@
 #define __al_included_allegro5_allegro_font_h
 
 #include "allegro5/allegro.h"
-#include "allegro5/internal/aintern_list.h"
 
 #if (defined ALLEGRO_MINGW32) || (defined ALLEGRO_MSVC) || (defined ALLEGRO_BCC32)
    #ifndef ALLEGRO_STATICLINK
@@ -47,19 +46,11 @@
 /* Type: ALLEGRO_FONT
 */
 typedef struct ALLEGRO_FONT ALLEGRO_FONT;
+#if defined(ALLEGRO_UNSTABLE) || defined(ALLEGRO_INTERNAL_UNSTABLE) || defined(ALLEGRO_FONT_SRC)
+
 /* Type: ALLEGRO_GLYPH
 */
 typedef struct ALLEGRO_GLYPH ALLEGRO_GLYPH;
-typedef struct ALLEGRO_FONT_VTABLE ALLEGRO_FONT_VTABLE;
-
-struct ALLEGRO_FONT
-{
-   void *data;
-   int height;
-   ALLEGRO_FONT *fallback;
-   ALLEGRO_FONT_VTABLE *vtable;
-   _AL_LIST_ITEM *dtor_item;
-};
 
 struct ALLEGRO_GLYPH
 {
@@ -73,30 +64,7 @@ struct ALLEGRO_GLYPH
    int offset_y;
    int advance;
 };
-
-/* text- and font-related stuff */
-struct ALLEGRO_FONT_VTABLE
-{
-   ALLEGRO_FONT_METHOD(int, font_height, (const ALLEGRO_FONT *f));
-   ALLEGRO_FONT_METHOD(int, font_ascent, (const ALLEGRO_FONT *f));
-   ALLEGRO_FONT_METHOD(int, font_descent, (const ALLEGRO_FONT *f));
-   ALLEGRO_FONT_METHOD(int, char_length, (const ALLEGRO_FONT *f, int ch));
-   ALLEGRO_FONT_METHOD(int, text_length, (const ALLEGRO_FONT *f, const ALLEGRO_USTR *text));
-   ALLEGRO_FONT_METHOD(int, render_char, (const ALLEGRO_FONT *f, ALLEGRO_COLOR color, int ch, float x, float y));
-   ALLEGRO_FONT_METHOD(int, render, (const ALLEGRO_FONT *f, ALLEGRO_COLOR color, const ALLEGRO_USTR *text, float x, float y));
-   ALLEGRO_FONT_METHOD(void, destroy, (ALLEGRO_FONT *f));
-   ALLEGRO_FONT_METHOD(void, get_text_dimensions, (const ALLEGRO_FONT *f,
-      const ALLEGRO_USTR *text, int *bbx, int *bby, int *bbw, int *bbh));
-   ALLEGRO_FONT_METHOD(int, get_font_ranges, (ALLEGRO_FONT *font,
-      int ranges_count, int *ranges));
-      
-   ALLEGRO_FONT_METHOD(bool, get_glyph_dimensions, (const ALLEGRO_FONT *f,
-      int codepoint, int *bbx, int *bby, int *bbw, int *bbh));      
-   ALLEGRO_FONT_METHOD(int, get_glyph_advance, (const ALLEGRO_FONT *font,
-      int codepoint1, int codepoint2));
-
-   ALLEGRO_FONT_METHOD(bool, get_glyph, (const ALLEGRO_FONT *f, int prev_codepoint, int codepoint, ALLEGRO_GLYPH *glyph));
-};
+#endif
 
 enum {
    ALLEGRO_NO_KERNING       = -1,
@@ -147,7 +115,9 @@ ALLEGRO_FONT_FUNC(bool, al_get_glyph_dimensions, (const ALLEGRO_FONT *f,
    int codepoint, int *bbx, int *bby, int *bbw, int *bbh));
 ALLEGRO_FONT_FUNC(int, al_get_glyph_advance, (const ALLEGRO_FONT *f,
    int codepoint1, int codepoint2));
+#if defined(ALLEGRO_UNSTABLE) || defined(ALLEGRO_INTERNAL_UNSTABLE) || defined(ALLEGRO_FONT_SRC)
 ALLEGRO_FONT_FUNC(bool, al_get_glyph, (const ALLEGRO_FONT *f, int prev_codepoint, int codepoint, ALLEGRO_GLYPH *glyph));
+#endif
 
 ALLEGRO_FONT_FUNC(void, al_draw_multiline_text, (const ALLEGRO_FONT *font, ALLEGRO_COLOR color, float x, float y, float max_width, float line_height, int flags, const char *text));
 ALLEGRO_FONT_FUNC(void, al_draw_multiline_textf, (const ALLEGRO_FONT *font, ALLEGRO_COLOR color, float x, float y, float max_width, float line_height, int flags, const char *format, ...));
